fix: address PR review findings — type safety, visual feedback, atomic cache
1. Narrow replyToMessageId from long? to int? at API boundary
   - IWebUserMessagingService, IWebBotMessagingService and implementations
   - Explicit (int) cast in Messages.HandleReplyToClick at model→API boundary
   - ChatInput.ReplyToMessageId parameter narrowed to int?
   - Full migration tracked in #214

2. Add StateHasChanged() for _actionInProgress visual feedback
   - Blazor only auto-renders after event handler completes
   - Without explicit StateHasChanged(), disabled state won't show during await
   - Applied to all 6 MudMenuItem handlers in MessageBubbleTelegram

3. Replace ConcurrentDictionary clear/refill with volatile reference swap
   - Old approach had a gap between Clear() and TryAdd() loop
   - New approach: build fresh Dictionary, atomic reference assignment
   - volatile ensures visibility across threads
   - Read-heavy/write-rare pattern — immutable snapshot is ideal

+++ TelegramGroupsAdmin.Telegram/Services/IWebUserMessagingService.cs
@@ -34,7 +34,7 @@ public interface IWebUserMessagingService
         WebUserIdentity webUser,
         long chatId,
         string text,
-        long? replyToMessageId = null,
+        int? replyToMessageId = null,
         CancellationToken cancellationToken = default);
 
     /// <summary>
